pyproject.toml
[project]
name = "fastapi-template"
version = "0.0.1"
description = "Template for FastAPI application"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.3",
    "bcrypt>=4.3.0",
    "fastapi>=0.116.0",
    "pydantic[email]>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "python-json-logger>=3.3.0",
    "redis>=6.2.0",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.35.0",
    "asyncpg>=0.30.0",
]

[dependency-groups]
dev = [
    "mypy>=1.16.1",
    "pre-commit>=4.2.0",
    "ruff>=0.12.4",
]
